% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_data}
\alias{sp_data}
\title{Financial variables for S&P prediction}
\format{
A data frame with 7944 rows and 24 variables

\describe{
\item{date}{dates}
\item{spx_index}{S&P 500 index, in points}
}
}
\usage{
sp_data
}
\description{
Financial variables for S&P prediction
}
\keyword{datasets}
